AASB_As_Above_So_Below/requirements.txt
# Data Manipulation and Analysis
pandas==1.5.3
numpy==1.24.3
dask==2023.10.1

# Statistical Modeling and Machine Learning
statsmodels==0.14.0
scikit-learn==1.2.2

# Visualization
plotly==5.15.0
dash==2.12.1
networkx==3.1

# Parallel Processing
joblib==1.3.2

# Configuration and Environment Management
PyYAML==6.0
python-dotenv==1.0.0

# Testing
pytest==7.2.2

# API Requests
requests==2.31.0

# Financial Data
yfinance==0.1.90
pytrends==4.10.0

# Jupyter Notebooks
jupyter==1.0.0
